refactor: упорядочил каталоги — удалил старый, новый переименовал в старое имя frontend
