+++ frontend/src/main.js
@@ -1,28 +1,54 @@
-import { el, mount } from "./utils/dom.js";
-import { TaskForm } from "./components/taskForm.js";
-import { TaskList } from "./components/taskList.js";
-import { enableAutoScrollZones } from "./features/autoscroll.js"; // ← НОВОЕ
+// склейка всего воедино
+import { createBus } from './core/bus.js';
+import { createStore } from './core/store.js';
+import { createRouter } from './core/router.js';
+import { loadConfig } from './core/config.js';
+import { createLayout } from './ui/Layout.js';
+import { enableButtonDelegation } from './ui/Button.js';
+import { createTodosView } from './features/todos/view/index.js';
+import { createTodoModel } from './features/todos/model.js';
 
-// подключаем стили динамически
-(() => {
-  const link = document.createElement("link");
-  link.rel = "stylesheet";
-  link.href = "src/styles.css";
-  document.head.append(link);
-})();
+const bus = createBus();
+const store = createStore({ todos: [] }, { key: 'app:v1' });
+const router = createRouter({ '/': true, '/about': true });
 
-(() => {
-  const header = el("header", {}, el("h1", { text: "ToDo List" }));
-  const main = el("main");
-  const footer = el("footer", {}, el("small", { text: "© 2025 ToDo App" }));
+const cfg = await loadConfig(); // { API_BASE: ... }
+console.log('Config:', cfg);
 
-  const list = TaskList();
-  const form = TaskForm(() => list._render?.());
+try {
+  console.log('before createLayout');
+  const layout = createLayout(); // layout живет внутри try
+  console.log('after createLayout', layout);
 
-  const listWrap = el("div", { className: "list-wrap" }, list);
-  mount(main, form, listWrap);
-  mount(document.body, header, main, footer);
+  const host = document.getElementById('app');
+  console.log('host #app =', host);
+  if (!host) throw new Error('No #app in DOM');
 
-  // Включаем зоны автопрокрутки именно на прокручиваемом контейнере
-  enableAutoScrollZones(listWrap, { zoneHeight: 44 });
-})();
+  layout.mountIn(host);
+  console.log('Layout mounted');
+
+  enableButtonDelegation(document);
+  console.log('Delegation enabled');
+
+  // Роутинг 
+  function renderRoute() {
+    const main = layout.getMain();
+    if (!main) throw new Error('layout.getMain() returned null/undefined');
+    main.replaceChildren();
+
+    const path = router.current();
+    if (path === '/') {
+      const model = createTodoModel(store);
+      const view = createTodosView(store, model);
+      main.append(view.el);
+    } else if (path === '/about') {
+      main.append(document.createTextNode('About page…'));
+    }
+  }
+
+  renderRoute();
+  router.onChange(renderRoute);
+
+} catch (e) {
+  console.error('BOOT ERROR:', e);
+}
