--- frontend/src/store/taskFactory.js
@@ -1,8 +1,0 @@
-export function createTask({ title }) {
-  return {
-    id: crypto.randomUUID(),
-    title: String(title).trim(),
-    done: false,
-    createdAt: new Date().toISOString(),
-  };
-}
