--- frontend/src/styles.css
@@ -1,301 +1,0 @@
-/* ================================
-   БАЗА И ПЕРЕМЕННЫЕ
-================================= */
-:root {
-  color-scheme: light dark;
-
-  --maxw: 900px;
-  --space-1: 4px;
-  --space-2: 8px;
-  --space-3: 12px;
-  --space-4: 16px;
-
-  --radius: 10px;
-  --border: 1px solid #d0d0d0;
-
-  --muted: #6b7280;
-  --danger: #b00020;
-  --accent: #3b82f6;
-}
-
-* { box-sizing: border-box; }
-
-/* ================================
-   ГЛОБАЛЬНАЯ ВЫСОТА/СКРОЛЛ
-================================= */
-html {
-  height: 100%;
-  overflow: hidden;
-  overscroll-behavior: none;
-}
-body {
-  position: fixed;
-  inset: 0;
-  margin: 0;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  font: 16px/1.4 system-ui, -apple-system, Segoe UI, Roboto, sans-serif;
-
-  overflow: hidden;
-  overscroll-behavior: none;
-  touch-action: none;
-}
-@supports (height: 100dvh) { body { height: 100dvh; } }
-
-/* во время DnD — запрет выделения текста и подсветки */
-.no-select, .no-select * {
-  -webkit-user-select: none;
-  user-select: none;
-  -webkit-tap-highlight-color: transparent;
-}
-
-/* ================================
-   КАРКАС
-================================= */
-header, footer {
-  flex: 0 0 auto;
-  width: 100%;
-  max-width: var(--maxw);
-  padding: var(--space-4);
-  margin: 0 auto;
-}
-main {
-  flex: 1 1 auto;
-  width: 100%;
-  max-width: var(--maxw);
-  padding: var(--space-4);
-  margin: 0 auto;
-
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-3);
-  min-height: 0;
-  overflow: hidden;
-}
-
-/* скроллим ТОЛЬКО список */
-.list-wrap {
-  flex: 1 1 auto;
-  min-height: 0;
-  overflow: auto;
-  padding-right: 4px;
-  -webkit-overflow-scrolling: touch;
-  overscroll-behavior: contain;
-  touch-action: pan-y;
-}
-.list-wrap.dragging { touch-action: none; }
-
-/* ================================
-   ФОРМА
-================================= */
-form {
-  display: flex;
-  gap: var(--space-2);
-  margin: var(--space-4) 0 var(--space-2);
-}
-input, button, select, textarea { font-size: 16px; }
-form input[type="text"] {
-  flex: 1;
-  padding: var(--space-3);
-  border: var(--border);
-  border-radius: var(--radius);
-  outline: none;
-}
-form button[type="submit"] {
-  padding: var(--space-3) var(--space-4);
-  border: var(--border);
-  border-radius: var(--radius);
-  background: transparent;
-  cursor: pointer;
-}
-
-/* ================================
-   СПИСОК ЗАДАЧ
-================================= */
-#task-list { list-style: none; padding: 0; margin: 0; }
-#task-list li {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  padding: var(--space-3) var(--space-4);
-  margin-bottom: var(--space-2);
-
-  border: var(--border);
-  border-radius: var(--radius);
-  background: transparent;
-
-  transition: transform 200ms ease, opacity 200ms ease, background-color 120ms ease;
-}
-#task-list li:hover { background: color-mix(in oklab, Canvas 90%, var(--accent) 10%); }
-#task-list li.dragging { opacity: 0.5; }
-
-/* промежуток вставки при DnD */
-.drop-placeholder {
-  height: 10px;
-  margin: var(--space-2) 0;
-  border-radius: var(--radius);
-  outline: 2px dashed var(--accent);
-  outline-offset: -4px;
-  background: linear-gradient(90deg, transparent, color-mix(in oklab, var(--accent) 20%, transparent), transparent);
-}
-
-/* ================================
-   КАРТОЧКИ ЗАДАЧ
-================================= */
-#task-list li.task-item {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  justify-content: initial;
-  padding: 0; /* внутренние паддинги у .task-row / .task-details */
-  will-change: clip-path, border-color;
-}
-
-/* верхняя строка плитки */
-.task-row {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  padding: var(--space-3) var(--space-4);
-}
-.task-title {
-  flex: 1 1 auto;
-  min-width: 0;
-  word-break: break-word;
-  overflow-wrap: anywhere;
-}
-.task-actions {
-  margin-left: auto;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-/* кнопки */
-.task-toggle {
-  border: var(--border);
-  background: transparent;
-  border-radius: var(--radius);
-  padding: 6px 12px;
-  font-size: 18px;
-  line-height: 1;
-  cursor: pointer;
-}
-button.delete {
-  border: 0;
-  background: transparent;
-  font-size: 18px;
-  line-height: 1;
-  cursor: pointer;
-  color: var(--danger);
-}
-
-/* раскрывающийся блок (анимация высоты — из JS/WAAPI) */
-.task-details {
-  overflow: clip;         /* не вылезать при анимации высоты */
-  height: 0;
-  opacity: 0;
-  padding: 0 var(--space-4);
-  border-top: 1px dashed #e0e0e0;
-
-  min-width: 0;
-  max-width: 100%;
-}
-.task-item.is-open .task-details { opacity: 1; }
-
-/* текст описания — переносы */
-.task-desc {
-  margin: 8px 0 12px;
-  color: inherit;
-  white-space: pre-wrap;
-  overflow-wrap: anywhere;
-  word-break: break-word;
-}
-
-/* действия внизу описания */
-.details-actions {
-  display: flex;
-  justify-content: flex-end;
-  padding: 8px var(--space-4) 4px;
-}
-.task-edit {
-  border: var(--border);
-  background: transparent;
-  border-radius: var(--radius);
-  padding: 4px 10px;
-  font-size: 14px;
-  cursor: pointer;
-}
-
-/* редактор */
-.task-editbar {
-  display: grid;
-  gap: 8px;
-  padding: 8px var(--space-4) 12px;
-}
-.edit-label { font-size: 12px; color: var(--muted); }
-.edit-title,
-.edit-desc {
-  width: 100%;
-  border: var(--border);
-  border-radius: var(--radius);
-  padding: 8px;
-  font: inherit;
-  resize: none;
-}
-.edit-actions {
-  display: flex;
-  gap: 8px;
-  justify-content: flex-end;
-  margin-top: 4px;
-}
-.task-save, .task-cancel {
-  border: var(--border);
-  background: transparent;
-  border-radius: var(--radius);
-  padding: 6px 10px;
-  cursor: pointer;
-}
-
-/* === Рамки/анимации === */
-
-/* Жёлтая при входе в редактирование: снизу вверх */
-#task-list li.task-item.editing-yellow { position: relative; }
-#task-list li.task-item.editing-yellow::after {
-  content: ""; position: absolute; inset: 0; border-radius: var(--radius);
-  pointer-events: none;
-  border: 2px solid #facc15;  /* amber-400 */
-  clip-path: inset(0 0 100% 0);
-  animation: border-fill-up 600ms ease forwards;
-}
-@keyframes border-fill-up { to { clip-path: inset(0 0 0 0); } }
-
-/* статичная жёлтая рамка, пока редактируем */
-#task-list li.task-item.editing-solid { border: 2px solid #facc15; }
-
-/* Зелёная при сохранении: оббегающая по периметру */
-#task-list li.task-item.saving-green { position: relative; }
-#task-list li.task-item.saving-green::after {
-  content: ""; position: absolute; inset: 0; border-radius: var(--radius);
-  pointer-events: none;
-  border: 2px solid #22c55e;  /* green-500 */
-  clip-path: inset(0 100% 0 0);
-  animation: border-sweep 900ms linear forwards;
-}
-@keyframes border-sweep {
-  0%   { clip-path: inset(0 100% 0 0); }
-  25%  { clip-path: inset(0 0 100% 0); }
-  50%  { clip-path: inset(100% 0 0 0); }
-  75%  { clip-path: inset(0 0 0 100%); }
-  100% { clip-path: inset(0 0 0 0); }
-}
-
-/* финал — белая рамка на короткое время */
-#task-list li.task-item.saved-white {
-  border: 2px solid #fff;
-  transition: border-color 300ms ease;
-}
